+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { TableConfiguration } from "./components/TableConfiguration";
-import { AddTrigger, CreateRow, Statistics } from "./pages";
+import { CreateRow, Statistics } from "./pages";
 import { SharedLayout } from "./components";
 import TriggerType from "./pages/AddTriggerType";
 
@@ -12,7 +12,6 @@ function App() {
         <Routes>
           <Route path="/" element={<SharedLayout />}>
             <Route path="/" element={<TableConfiguration />} />
-            <Route path="/add_trigger" element={<AddTrigger />} />
             <Route path="/statistics" element={<Statistics />} />
             <Route path="/data-form" element={<CreateRow />} />
             <Route path="/trigger-type" element={<TriggerType />} />
