+++ client/src/components/ParameterTable.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useQuery } from "react-query";
 import { baseURL } from "../constants/urlConstants";
-import TableLookUps from "../constants/TableLookUps";
+import translate from "../services/translate";
 
 const fetchData = async (url) => {
   const res = await fetch(url);
@@ -29,6 +29,8 @@ const ParameterTable = () => {
     }
   };
 
+  const language = localStorage.getItem("language");
+
   const handleAddRange = (rowId) => {
     // Implement your logic to add a new range to the parameter
     console.log(`Add range to Parameter ${rowId}`);
@@ -48,9 +50,9 @@ const ParameterTable = () => {
     <table>
       <thead>
         <tr>
-          <th>{TableLookUps("Name")}</th>
-          <th>{TableLookUps("Unit")}</th>
-          <th>{TableLookUps("Ranges")}</th>
+          <th>{translate(language, "Name")}</th>
+          <th>{translate(language, "Unit")}</th>
+          <th>{translate(language, "Ranges")}</th>
         </tr>
       </thead>
       <tbody>
@@ -62,7 +64,7 @@ const ParameterTable = () => {
                 <td>{parameter.unit}</td>
                 <td>
                   <button onClick={() => handleRowExpand(parameter._id)}>
-                    {TableLookUps("Expand")}
+                    {translate(language, "Expand")}
                   </button>
                 </td>
               </tr>
@@ -72,10 +74,10 @@ const ParameterTable = () => {
                     <table>
                       <thead>
                         <tr>
-                          <th>{TableLookUps("Message")}</th>
-                          <th>{TableLookUps("Min")}</th>
-                          <th>{TableLookUps("Max")}</th>
-                          <th>{TableLookUps("Actions")}</th>
+                          <th>{translate(language, "Message")}</th>
+                          <th>{translate(language, "Min")}</th>
+                          <th>{translate(language, "Max")}</th>
+                          <th>{translate(language, "Actions")}</th>
                         </tr>
                       </thead>
                       <tbody>
@@ -90,14 +92,14 @@ const ParameterTable = () => {
                                   handleUpdateRange(parameter._id, range._id)
                                 }
                               >
-                                {TableLookUps("Update")}
+                                {translate(language, "Update")}
                               </button>
                               <button
                                 onClick={() =>
                                   handleDeleteRange(parameter._id, range._id)
                                 }
                               >
-                                {TableLookUps("Delete")}
+                                {translate(language, "Delete")}
                               </button>
                             </td>
                           </tr>
@@ -116,7 +118,7 @@ const ParameterTable = () => {
                             <button
                               onClick={() => handleAddRange(parameter._id)}
                             >
-                              {TableLookUps("Add Range")}
+                              {translate(language, "Add Range")}
                             </button>
                           </td>
                         </tr>
