+++ client/src/components/SensorsTable.jsx
@@ -1,67 +1,68 @@
 import React, { useEffect, useState } from "react";
 import { Tables } from "../CustomHook/Tables.Custom";
-import TableLookUps from "../constants/TableLookUps";
+import translate from "../services/translate";
 
 const SensorsTable = () => {
   const [rows, setRows] = useState([]);
+  const language = localStorage.getItem("language");
   const sensorsColumns = [
     { field: "id", headerName: "Sensor ID", flex: 1, hide: true },
     //trigger number
     {
       field: "trigger",
-      headerName: TableLookUps("trigger"),
+      headerName: translate(language, "trigger"),
       flex: 1,
     },
     //trigger name
     {
       field: "triggerName",
-      headerName: TableLookUps("triggerName"),
+      headerName: translate(language, "triggerName"),
       flex: 1,
     },
     //trigger number
 
     {
       field: "triggerType",
-      headerName: TableLookUps("triggerType"),
+      headerName: translate(language, "triggerType"),
       flex: 1,
     },
     //Sensore
     //sensorName
     {
       field: "sensorName",
-      headerName: TableLookUps("sensorName"),
+      headerName: translate(language, "sensorName"),
       flex: 1,
     },
     //Sensore
     //sensorType
     {
       field: "sensorType",
-      headerName: TableLookUps("sensorType"),
+      headerName: translate(language, "sensorType"),
       flex: 1,
     },
     //Sensore
     //System
     {
       field: "System",
-      headerName: TableLookUps("System"),
+      headerName: translate(language, "System"),
       flex: 1,
     },
     //Sensore
     //SystemNumber
     {
       field: "SystemNumber",
-      headerName: TableLookUps("SystemNumber"),
+      headerName: translate(language, "SystemNumber"),
       flex: 1,
     },
     {
       field: "createdAt",
-      headerName: TableLookUps("createdAt"),
+      headerName: translate(language, "createdAt"),
 
       flex: 1,
     },
     {
       field: "updatedAt",
-      headerName: TableLookUps("updatedAt"),
+      headerName: translate(language, "updatedAt"),
 
       flex: 1,
     },
@@ -87,7 +88,7 @@ const SensorsTable = () => {
 
   return (
     <div>
-      <h2>{TableLookUps("sensors")}</h2>
+      <h2>{translate(language, "sensors")}</h2>
       <Tables
         QueryName={"Sensors"}
         Massage={massage}
