--- client/src/constants/TableLookUps.js
@@ -1,155 +1,0 @@
-export default (word) => {
-  //get the lang out of attribute as has been set in the constants/index.js file
-  let lang = document.getElementsByTagName("html")[0].getAttribute("lang");
-
-  const lookUps = {
-    he: {
-      triggers: "טריגרים",
-      "Config Conditions": "תנאי הגדרות ",
-      city: "עיר",
-      street: "רחוב ",
-      number: "מספר",
-      latitude: "קו רוחב",
-      longitude: "קו אורך",
-      createdAt: "נוצר בתאריך",
-      updatedAt: "עודכן בתאריך",
-      IdentifierID: "זיהוי",
-      trigger: "טריגר",
-      meaning: "טריגר ",
-      number: "מספר",
-      system: "מערכת",
-      status: "מעמד",
-      additionalIdentifier: "זיהוי נוסף",
-      status: "מצב",
-      "config Condition ID": "מזהה תנאי הגדרות",
-      systemId: "מזהה המערכת",
-      previousSystemId: "מזהה המערכת הקודמת",
-      type: "סוג",
-      triggerName: "שם הטריגר",
-      valueType: "סוג הערך",
-      measuredValue: "ערך מודד",
-      Level: "רמה",
-      Identifiers: "מזהים",
-      Sensors: "חיישנים ",
-      SystemNumber: "מספר המערכת",
-      Actions: "פעולות",
-      Parameters: "פרמטרים",
-      Activity: "פעילות",
-      Meaning: "משמעות",
-      Name: "שם",
-      Unit: "יחידה",
-      Ranges: "טווחים",
-      "Checked Items": "פריטים שנבדקו",
-      Expand: "להרחיב",
-      Message: "הודעה",
-      Min: "מינימום",
-      Max: "מקסימום",
-      Actions: "פעולות",
-      edit: "לערוך",
-      "Add Range": "הוסף טווח",
-      Delete: "למחוק",
-    },
-    en: {
-      Delete: "Delete",
-      "Add Range": "Add Range",
-      edit: "Edit",
-      Message: "Message",
-      Min: "Min",
-      Max: "Max",
-      Actions: "Actions",
-      Expand: "Expand",
-      "Checked Items": "Checked Items",
-      Ranges: "Ranges",
-      Unit: "Unit",
-      Name: "Name",
-      Meaning: "Meaning",
-      Activity: "Activity",
-      Parameters: "Parameters",
-      Actions: "Actions",
-      SystemNumber: "System Number",
-      Sensors: "Sensors",
-      Identifiers: "Identifiers",
-      triggers: "Triggers",
-      "Config Conditions": "Config Conditions",
-      Level: "Level",
-      valueType: "value Type",
-      measuredValue: "Measured Value",
-      triggerType:"Trigger Type",
-      sensorName:"Sensor Name",
-      sensorType:"Sensor Type",
-      SystemNumber:"System Number",
-      triggerName: "Trigger Name",
-      System:"System Type",
-      type: "type",
-      previousSystemId: "Previous System Id",
-      systemId: "System ID",
-      "config Condition ID": "config Condition ID",
-      status: "Status",
-      city: "City",
-      street: "Street",
-      number: "Number",
-      latitude: "Latitude",
-      longitude: "Longitude",
-      createdAt: "Created",
-      updatedAt: "Updated",
-      IdentifierID: "Identifier ID",
-      additionalIdentifier: "Additional Identifier",
-      trigger: "Trigger",
-      meaning: "Meaning",
-      number: "Number(Sensor)",
-      system: "System",
-      status: "Status",
-    },
-    ar: {
-      Delete: "حذف",
-      "Add Range": "اضافة حد",
-      edit: "تعديل",
-      Message: "رسالة",
-      Min: "حد ادنى",
-      Max: "حد اعلى",
-      Actions: "إجراءات",
-      Expand: "تضخيم",
-      "Checked Items": "العناصر المختارة",
-      Ranges: "حدود",
-      Unit: "وحدة",
-      Name: "اسم",
-      Meaning: "معنى",
-      Activity: "فعليات",
-      Parameters: "المعاملات",
-      SystemNumber: "رقم النظام",
-      Sensors: "حساسات",
-      Identifiers: "المعرفين",
-      triggers: "المؤثرات",
-      "Config Conditions": " شروط التكوين ",
-      Level: "مستوى",
-      valueType: "نوع القيمة",
-      measuredValue: "القيمة المحسوبة",
-      triggerName: "اسم المؤثر",
-      type: "نوع",
-      previousSystemId: "رقم النظام السابق",
-      systemId: "رقم النظام",
-      "config Condition ID": "معرف شرط التكوين",
-      status: "الوضع",
-      city: "المدينة",
-      street: "الشارع",
-      number: "الرقم",
-      latitude: "خطوط الطول",
-      longitude: "خطوط العرض",
-      createdAt: "تم إنشاؤه",
-      updatedAt: "تم تعديله",
-      IdentifierID: "رقم المعرف",
-      additionalIdentifier: "معرف اضافي",
-      trigger: "المؤثر",
-      meaning: "المعنى",
-      number: "رقم",
-      system: "النظام",
-    },
-  };
-  ////if set lang is not supported do back to english///
-  if (!lookUps[lang]) lang = "en";
-
-  if (!lookUps[lang][word]) {
-    return "translation error";
-  }
-  return lookUps[lang][word];
-};
