--- client/src/constants/englishLanguage.js
@@ -1,31 +1,0 @@
-const  ENGLISH_LANGUAGE = {
-    TRIGGER_FIELD_ERROR :  "-",
-    NUMBER_FIELD_ERROR : "-",
-    MEANING_FIELD_ERROR : "-",
-    SYSTEM_FIELD_ERROR : "-",
-    SUCCESS_ADD_ROW_MSG : "The data sent successfully",
-    WAITING_BUTTON : "Waiting...",
-    SEND_TRIGGER_BUTTON : "Send trigger",
-    ADD_TRIGGER_TITLE : "Add a new trigger",
-    NUMBER_FIELD : "Number",
-    TRIGGER_FIELD : "Trigger",
-    SYSTEM_FIELD : "System",
-    MEANING_FIELD : "Meaning",
-    LANGUAGE_BUTTON : "Change language",
-    STATISTIC_BUTTON : "Show handled inquiries:",
-    STATISTIC_TYPE_A_OPTION : "Daily",
-    STATISTIC_TYPE_B_OPTION : "Weekly",
-    STATISTIC_TYPE_C_OPTION : "Monthly",
-    STATISTIC_TYPE_D_OPTION : "Yearly",
-    OPEN_INQUIRIES_LABEL : "Waiting for treatment",
-    HANDLED_ALERT_LABEL : "Closed inquiries",
-    CHART_OPTIONS_TEXT : "Status of inquiries",
-    CHART_TITLE_H2 : "Inquiries handled",
-    TOTAL_LABEL : "Total",
-    LOADING_ALT : "LOADING...",
-
-
-}
-
-
-export default ENGLISH_LANGUAGE;
