--- client/src/constants/hebrewLanguage.js
@@ -1,33 +1,0 @@
-const  HEBREW_LANGUAGE = {
-    TRIGGER_FIELD_ERROR :  "-",
-    NUMBER_FIELD_ERROR : "-",
-    MEANING_FIELD_ERROR : "-",
-    SYSTEM_FIELD_ERROR : "-",
-    SUCCESS_ADD_ROW_MSG : 'הנתונים נשלחו בהצלחה',
-    WAITING_BUTTON : "המתן...",
-    SEND_TRIGGER_BUTTON : "שלח טריגר",
-    ADD_TRIGGER_TITLE : " הוספת טריגר",
-    NUMBER_FIELD : "מספר",
-    TRIGGER_FIELD : "טריגר",
-    SYSTEM_FIELD : "מערכת",
-    MEANING_FIELD : "משמעות",
-    LANGUAGE_BUTTON : "שינוי שפה",
-    STATISTIC_BUTTON : "בחר סטטיסטיקת פניות להצגה",
-    STATISTIC_TYPE_A_OPTION : "יומית",
-    STATISTIC_TYPE_B_OPTION : "שבועית",
-    STATISTIC_TYPE_C_OPTION : "חודשית",
-    STATISTIC_TYPE_D_OPTION : "שנתית",
-    OPEN_INQUIRIES_LABEL : "ממתינות לטיפול",
-    HANDLED_ALERT_LABEL : "טופלו",
-    CHART_OPTIONS_TEXT : "סטטוס פניות",
-    CHART_TITLE_H2 : "פניות שטופלו",
-    TOTAL_LABEL : 'סה"כ',
-    LOADING_ALT : "טוען...",
-
-
-
-
-
-}
-
-export default HEBREW_LANGUAGE;
