+++ client/src/constants/index.js
@@ -1,9 +1,5 @@
-import  HEBREW_LANGUAGE  from "./hebrewLanguage";
-import  ENGLISH_LANGUAGE  from "./englishLanguage";
-import  {getItem} from "../services/localStorageService";
-
 export const DEFAULT_LANGUAGE = "he";
-export let language = getItem('language') || DEFAULT_LANGUAGE;//set hebrew language by default
+export let language = localStorage.getItem("language") || DEFAULT_LANGUAGE; //set hebrew language by default
 export const BASE_URL_SERVER = "http://localhost:5000";
 export const MS_1000 = 1000;
 export const DEFAULT_TIME_ALERT = 3;
@@ -13,106 +9,3 @@ export const DANGER_TIME_ALERT = 2.5;
 export const MIN_LEN_NUMBER = 2;
 export const MIN_LEN_MEANING = 2;
 export const SUCCESS = true;
-
-export let TRIGGER_FIELD_ERROR ;
-export let NUMBER_FIELD_ERROR ;
-export let MEANING_FIELD_ERROR ;
-export let SYSTEM_FIELD_ERROR ;
-export let SUCCESS_ADD_ROW_MSG ;
-export let WAITING_BUTTON ;
-export let SEND_TRIGGER_BUTTON ;
-export let ADD_TRIGGER_TITLE ;
-export let NUMBER_FIELD ;
-export let TRIGGER_FIELD ;
-export let SYSTEM_FIELD ;
-export let MEANING_FIELD ;
-export let LANGUAGE_BUTTON ;
-export let STATISTIC_BUTTON ;
-export let STATISTIC_TYPE_A_OPTION ;
-export let STATISTIC_TYPE_B_OPTION ;
-export let STATISTIC_TYPE_C_OPTION ;
-export let STATISTIC_TYPE_D_OPTION ;
-export let OPEN_INQUIRIES_LABEL;
-export let HANDLED_ALERT_LABEL;
-export let TOTAL_LABEL;
-export let CHART_OPTIONS_TEXT;
-export let CHART_TITLE_H2;
-export let LOADING_ALT;
-
-
-
-
-export const setLanguage = (lang = "he") => {
-
-    if (lang === "he")
-    {
-        
-        TRIGGER_FIELD_ERROR = HEBREW_LANGUAGE.TRIGGER_FIELD_ERROR;
-        NUMBER_FIELD_ERROR =  HEBREW_LANGUAGE.NUMBER_FIELD_ERROR;
-        MEANING_FIELD_ERROR = HEBREW_LANGUAGE.MEANING_FIELD_ERROR;
-        SYSTEM_FIELD_ERROR = HEBREW_LANGUAGE.SYSTEM_FIELD_ERROR;
-        SUCCESS_ADD_ROW_MSG = HEBREW_LANGUAGE.SUCCESS_ADD_ROW_MSG;
-        WAITING_BUTTON = HEBREW_LANGUAGE.WAITING_BUTTON;
-        SEND_TRIGGER_BUTTON = HEBREW_LANGUAGE.SEND_TRIGGER_BUTTON;
-        ADD_TRIGGER_TITLE = HEBREW_LANGUAGE.ADD_TRIGGER_TITLE;
-        NUMBER_FIELD = HEBREW_LANGUAGE.NUMBER_FIELD;
-        TRIGGER_FIELD = HEBREW_LANGUAGE.TRIGGER_FIELD;
-        SYSTEM_FIELD = HEBREW_LANGUAGE.SYSTEM_FIELD;
-        MEANING_FIELD = HEBREW_LANGUAGE.MEANING_FIELD;
-        LANGUAGE_BUTTON = HEBREW_LANGUAGE.LANGUAGE_BUTTON;
-        STATISTIC_BUTTON = HEBREW_LANGUAGE.STATISTIC_BUTTON;
-        STATISTIC_TYPE_A_OPTION = HEBREW_LANGUAGE.STATISTIC_TYPE_A_OPTION;
-        STATISTIC_TYPE_B_OPTION = HEBREW_LANGUAGE.STATISTIC_TYPE_B_OPTION;
-        STATISTIC_TYPE_C_OPTION = HEBREW_LANGUAGE.STATISTIC_TYPE_C_OPTION;
-        STATISTIC_TYPE_D_OPTION = HEBREW_LANGUAGE.STATISTIC_TYPE_D_OPTION;
-        OPEN_INQUIRIES_LABEL = HEBREW_LANGUAGE.OPEN_INQUIRIES_LABEL;
-        HANDLED_ALERT_LABEL = HEBREW_LANGUAGE.HANDLED_ALERT_LABEL; 
-        CHART_OPTIONS_TEXT = HEBREW_LANGUAGE.CHART_OPTIONS_TEXT; 
-        CHART_TITLE_H2 = HEBREW_LANGUAGE.CHART_TITLE_H2;
-        TOTAL_LABEL = HEBREW_LANGUAGE.TOTAL_LABEL;
-        LOADING_ALT = HEBREW_LANGUAGE.LOADING_ALT; 
-        document.getElementsByTagName('html')[0].setAttribute("dir", "rtl");
-
-    }
-    else if (lang === "en") {
-        TRIGGER_FIELD_ERROR = ENGLISH_LANGUAGE.TRIGGER_FIELD_ERROR;
-        NUMBER_FIELD_ERROR =  ENGLISH_LANGUAGE.NUMBER_FIELD_ERROR;
-        MEANING_FIELD_ERROR = ENGLISH_LANGUAGE.MEANING_FIELD_ERROR;
-        SYSTEM_FIELD_ERROR = ENGLISH_LANGUAGE.SYSTEM_FIELD_ERROR;
-        SUCCESS_ADD_ROW_MSG = ENGLISH_LANGUAGE.SUCCESS_ADD_ROW_MSG;
-        WAITING_BUTTON = ENGLISH_LANGUAGE.WAITING_BUTTON;
-        SEND_TRIGGER_BUTTON = ENGLISH_LANGUAGE.SEND_TRIGGER_BUTTON;
-        ADD_TRIGGER_TITLE = ENGLISH_LANGUAGE.ADD_TRIGGER_TITLE;
-        NUMBER_FIELD = ENGLISH_LANGUAGE.NUMBER_FIELD;
-        TRIGGER_FIELD = ENGLISH_LANGUAGE.TRIGGER_FIELD;
-        SYSTEM_FIELD = ENGLISH_LANGUAGE.SYSTEM_FIELD;
-        MEANING_FIELD = ENGLISH_LANGUAGE.MEANING_FIELD;
-        LANGUAGE_BUTTON = ENGLISH_LANGUAGE.LANGUAGE_BUTTON;
-        STATISTIC_BUTTON = ENGLISH_LANGUAGE.STATISTIC_BUTTON;
-        STATISTIC_TYPE_A_OPTION = ENGLISH_LANGUAGE.STATISTIC_TYPE_A_OPTION;
-        STATISTIC_TYPE_B_OPTION = ENGLISH_LANGUAGE.STATISTIC_TYPE_B_OPTION;
-        STATISTIC_TYPE_C_OPTION = ENGLISH_LANGUAGE.STATISTIC_TYPE_C_OPTION;
-        STATISTIC_TYPE_D_OPTION = ENGLISH_LANGUAGE.STATISTIC_TYPE_D_OPTION;
-        OPEN_INQUIRIES_LABEL = ENGLISH_LANGUAGE.OPEN_INQUIRIES_LABEL;
-        HANDLED_ALERT_LABEL = ENGLISH_LANGUAGE.HANDLED_ALERT_LABEL;
-        CHART_OPTIONS_TEXT = ENGLISH_LANGUAGE.CHART_OPTIONS_TEXT; 
-        CHART_TITLE_H2 = ENGLISH_LANGUAGE.CHART_TITLE_H2; 
-        TOTAL_LABEL = ENGLISH_LANGUAGE.TOTAL_LABEL;
-        LOADING_ALT = ENGLISH_LANGUAGE.LOADING_ALT; 
- 
-        document.getElementsByTagName('html')[0].setAttribute("dir", "ltr");
-    }
-    document.getElementsByTagName('html')[0].setAttribute("lang", lang);
-     language = lang
-
-}
-setLanguage(language);
-
-
-
-
-
-
-
-
-
