+++ client/src/pages/Statistics.page.jsx
@@ -1,24 +1,17 @@
-
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import useChoseTypeStatistic from "../CustomHook/useChoseTypeStatistic";
-import {Statistic} from "../components";
-import {
-  STATISTIC_BUTTON,
-  STATISTIC_TYPE_A_OPTION,
-  STATISTIC_TYPE_B_OPTION,
-  STATISTIC_TYPE_C_OPTION,
-  STATISTIC_TYPE_D_OPTION
-} from "../constants"
+import { Statistic } from "../components";
+import translate from "../services/translate";
+
 const Statistics = () => {
   const [chooseType, setChooseType] = useState(null);
-  const { statisticType,
-  } = useChoseTypeStatistic();
-  useEffect(() => {
-  }, []);
+  const { statisticType } = useChoseTypeStatistic();
+
+  const language = localStorage.getItem("language");
 
   const setStatisticType = (e) => {
-    setChooseType(e.target.value)
-    };
+    setChooseType(e.target.value);
+  };
 
   return (
     <>
@@ -29,14 +22,14 @@ const Statistics = () => {
         onChange={setStatisticType}
         className={`form-input m-t`}
       >
-        <option value=''>{STATISTIC_BUTTON}</option>
-        <option value="day">{STATISTIC_TYPE_A_OPTION}</option>
-        <option value="week">{STATISTIC_TYPE_B_OPTION}</option>
-        <option value="month">{STATISTIC_TYPE_C_OPTION}</option>
-        <option value="year">{STATISTIC_TYPE_D_OPTION}</option>
+        <option value="">{translate(language, "")}</option>
+        <option value="day">{translate(language, "Daily")}</option>
+        <option value="week">{translate(language, "Weekly")}</option>
+        <option value="month">{translate(language, "Monthly")}</option>
+        <option value="year">{translate(language, "Yearly")}</option>
       </select>
-{     chooseType &&  <Statistic statisticType={chooseType} />
-}    </>
+      {chooseType && <Statistic statisticType={chooseType} />}{" "}
+    </>
   );
 };
