+++ client/src/pages/index.js
@@ -1,5 +1,4 @@
-import AddTrigger from './AddTrigger.page';
-import CreateRow from "./CreateRow.page"
-import Statistics from './Statistics.page';
+import CreateRow from "./CreateRow.page";
+import Statistics from "./Statistics.page";
 
-export { AddTrigger, Statistics, CreateRow };
+export { Statistics, CreateRow };
